pyproject.toml
[tool.poetry]
name = "distributed-microservices"
version = "0.1.0"
description = ""
authors = ["Tega Ukavwe <justtega97@gmail.com>"]
readme = "README.md"
packages = [{include = "distributed_microservices"}]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
encrypted-mysqldb = "^1.0.0"
privex-pyjwt = "^2.0.0"
flask-mysqldb = "^1.0.1"
python-dotenv = "^0.21.0"
cryptography = "3.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
